+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 New features, fixed bugs, known defects and other noteworthy changes to each release of the Catena-X Portal Frontend.
 
 ### Unreleased
+* Feature - Design/BL&API for display app details for Admin Board
 
 ## 1.0.0-RC6